+++ libp2p/src/builder/phase/other_transport.rs
@@ -95,6 +95,22 @@ impl<T: AuthenticatedMultiplexedTransport>
         self.without_any_other_transports()
             .with_dns_config(cfg, opts)
     }
+
+    /// See [`SwarmBuilder::try_with_dns_config`].
+    pub fn try_with_dns_config(
+        self,
+        cfg: libp2p_dns::ResolverConfig,
+        opts: libp2p_dns::ResolverOpts,
+    ) -> Result<
+        SwarmBuilder<
+            super::provider::Tokio,
+            WebsocketPhase<impl AuthenticatedMultiplexedTransport>,
+        >,
+        libp2p_dns::ResolveError,
+    > {
+        self.without_any_other_transports()
+            .try_with_dns_config(cfg, opts)
+    }
 }
 #[cfg(feature = "relay")]
 impl<T: AuthenticatedMultiplexedTransport, Provider>