+++ transports/dns/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.45.0
 
+- Add `tokio::Transport::try_custom` for fallible custom resolver construction.
+  See [PR 6418](https://github.com/libp2p/rust-libp2p/pull/6418).
+
 - Update DNS transport internals to support `hickory-resolver` 0.26.
   See [PR 6418](https://github.com/libp2p/rust-libp2p/pull/6418).